requirements.txt
requests
openai
beautifulsoup4
python-dotenv
markdown
weasyprint
html-sanitizer